chore(core): extract nx migrate helpers into sibling modules
Split helpers introduced/extended by the recent --mode work out of the
3k-line migrate.ts into focused sibling files. Public-export surface
(Migrator, parseMigrationsOptions, executeMigrations, normalizeVersion,
filterDowngradedUpdates, etc.) is preserved, including the import paths
used by migrate-ui-api, repair, and migrate.spec.

- version-utils.ts: shared atoms with no migrate-specific deps
  (DIST_TAGS, normalizeVersion, normalizeVersionWithTagCheck,
  isNxEquivalentTarget). Imported by the other three sibling files;
  none of them import from migrate.ts (acyclic graph).
- multi-major.ts: stepwise-migration prompt/warn helpers, exposing
  only maybePromptOrWarnMultiMajorMigration to migrate.ts.
- update-filters.ts: filterDowngradedUpdates, re-exported from
  migrate.ts so existing test/external imports keep working.
- parseMigrationsOptions split into three named private helpers
  (assertThirdPartyModeFlagCompatibility, resolveTargetAndMode,
  assertThirdPartyTargetBounds). Public signature unchanged.
